requirements.txt
yfinance
pandas
dwave-ocean-sdk
qiskit
gurobipy
qiskit-optimization
